+++ src/ucsf-vpn.sh
@@ -79,7 +79,7 @@
 ### * UCSF Managing Your Passwords:
 ###   - https://it.ucsf.edu/services/managing-your-passwords
 ###
-### Version: 6.9.9-9001
+### Version: 6.9.9-9002
 ### Copyright: Henrik Bengtsson (2016-2025)
 ### License: GPL (>= 2.1) [https://www.gnu.org/licenses/gpl.html]
 ### Source: https://github.com/HenrikBengtsson/ucsf-vpn
@@ -184,7 +184,7 @@ function status() {
             else
                 connected+=(false)
             fi
-            msgs+=("Public IP information (ipinfo.im): $(public_info)")
+            msgs+=("Public IP information: $(public_info)")
         else
             merror "Unknown --validate value: $method"
         fi