+++ NEWS.md
@@ -8,8 +8,8 @@ ucsf-vpn
  * `ucsf-vpn' supports the GlobalProtect VPN protocol that UCSF
    migrated to on 2025-04-16.
 
- * `uscf-vpn details` queries <https://ipinfo.im/json>, because
-   <https://ipinfo.io/> is blocked on the UCSF network.
+ * `uscf-vpn details` queries the <https://ip-api.com/> service,
+   because 'ipinfo.io' is blocked on the UCSF network.
 
 
 ## Version 6.3.0 (2025-11-13)