+++ NEWS.md
@@ -1,7 +1,7 @@
 ucsf-vpn
 ========
 
-## Version (development version)
+## Version 7.0.0 (2025-12-08)
 
 ### Significant changes